Refactor to make __main__.py shorter.
